ComponentTab now edits TransformComponent

+++ src/Game/Components/TransformComponent.h
@@ -6,9 +6,9 @@
 struct TransformComponent {
     glm::vec2 position;
     glm::vec2 scale;
-    double rotation;
+    float rotation;
 
-    TransformComponent(glm::vec2 position = glm::vec2(0,0), glm::vec2 scale = glm::vec2(1, 1), double rotation = 0.0)
+    TransformComponent(glm::vec2 position = glm::vec2(0,0), glm::vec2 scale = glm::vec2(1, 1), float rotation = 0.0)
     {
         this->position = position;
         this->scale = scale;
